Combine SwiftUI demo to the main iOS demo
