+++ Demo/Demo/Kingfisher-Demo/SwiftUIViews/SwiftUIView.swift
@@ -1,8 +1,8 @@
 //
-//  MainView.swift
+//  SwiftUIView.swift
 //  Kingfisher
 //
-//  Created by jp20028 on 2019/08/07.
+//  Created by Wei Wang on 2019/06/18.
 //
 //  Copyright (c) 2019 Wei Wang <onevcat@gmail.com>
 //
@@ -24,34 +24,44 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
 
+import Kingfisher
 import SwiftUI
-import class Kingfisher.KingfisherManager
 
-struct MainView: View {
+@available(iOS 13.0, *)
+struct SwiftUIView : View {
+
+    @State private var index = 1
+
     var body: some View {
+        VStack {
+            KFImage(URL(string: "https://raw.githubusercontent.com/onevcat/Kingfisher-TestImages" +
+                                "/master/DemoAppImage/Loading/kingfisher-\(self.index).jpg")!)
+                .onSuccess { r in
+                    print("suc: \(r)")
+                }
+                .onFailure { e in
+                    print("err: \(e)")
+                }
+                .placeholder {
+                    Image(systemName: "arrow.2.circlepath.circle")
+                        .font(.largeTitle)
+                }
+                .resizable()
+                .frame(width: 300, height: 300)
+                .cornerRadius(20)
+                .shadow(radius: 5)
 
-        NavigationView {
-            List {
-                Button(
-                    action: {
-                        KingfisherManager.shared.cache.clearMemoryCache()
-                        KingfisherManager.shared.cache.clearDiskCache()
-                    },
-                    label: {
-                        Text("Clear Cache").foregroundColor(.blue)
-                    }
-                )
-                NavigationLink(destination: SwiftUIView()) { Text("Basic Image") }
-                NavigationLink(destination: SwiftUIList()) { Text("List") }
-            }.navigationBarTitle(Text("Kingfisher"))
-        }
+            Button(action: {
+                self.index = (self.index % 10) + 1
+            }) { Text("Next Image") }
+
+        }.navigationBarTitle(Text("Basic Image"), displayMode: .inline)
     }
 }
 
-#if DEBUG
-struct MainView_Previews: PreviewProvider {
+@available(iOS 13.0, *)
+struct SwiftUIView_Previews : PreviewProvider {
     static var previews: some View {
-        MainView()
+        SwiftUIView()
     }
 }
-#endif
